GlobeSpotter for ArcGIS Pro: - Version 0.8.2:   - Measurements:     - Point, line and surface measurements     - Smart click and standard measurements     - Cyclorama and map measurements

+++ client-side/C#/GlobeSpotterArcGISPro/GlobeSpotterArcGISPro/Configuration/File/ConstantsViewer.cs
@@ -78,7 +78,7 @@ namespace GlobeSpotterArcGISPro.Configuration.File
     /// <summary>
     /// Size of the measurement font
     /// </summary>
-    public double MeasurementFontSize { get; set; }
+    public float MeasurementFontSize { get; set; }
 
     /// <summary>
     /// MeasurementName
@@ -133,7 +133,7 @@ namespace GlobeSpotterArcGISPro.Configuration.File
         AddressLanguageCode = "nl",
         CrossCheckSize = 10.0,
         MeasurementPointSize = 5.0,
-        MeasurementFontSize = 8.0,
+        MeasurementFontSize = 8.0f,
         MeasurementName = "my measurement"
       };
